+++ proxy/src/http/source.rs
@@ -13,7 +13,7 @@ use crate::identity;
 
 /// Prefixed filters.
 pub fn routes(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("source").and(
         blob_filter(Arc::clone(&peer))
@@ -30,7 +30,7 @@ pub fn routes(
 /// Combination of all source filters.
 #[cfg(test)]
 fn filters(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     blob_filter(Arc::clone(&peer))
         .or(branches_filter(Arc::clone(&peer)))
@@ -44,7 +44,7 @@ fn filters(
 
 /// `GET /blob/<project_id>?revision=<revision>&path=<path>`
 fn blob_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("blob")
         .and(warp::get())
@@ -75,7 +75,7 @@ fn blob_filter(
 
 /// `GET /branches/<project_id>`
 fn branches_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("branches")
         .and(warp::get())
@@ -101,7 +101,7 @@ fn branches_filter(
 
 /// `GET /commit/<project_id>/<sha1>`
 fn commit_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("commit")
         .and(warp::get())
@@ -125,7 +125,7 @@ fn commit_filter(
 
 /// `GET /commits/<project_id>?branch=<branch>`
 fn commits_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("commits")
         .and(warp::get())
@@ -179,7 +179,7 @@ fn local_state_filter() -> impl Filter<Extract = impl Reply, Error = Rejection> 
 
 /// `GET /revisions/<project_id>`
 fn revisions_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("revisions")
         .and(warp::get())
@@ -207,7 +207,7 @@ fn revisions_filter(
 
 /// `GET /tags/<project_id>`
 fn tags_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("tags")
         .and(warp::get())
@@ -231,7 +231,7 @@ fn tags_filter(
 
 /// `GET /tree/<project_id>/<revision>/<prefix>`
 fn tree_filter(
-    peer: Arc<Mutex<coco::Peer>>,
+    peer: Arc<Mutex<coco::PeerApi>>,
 ) -> impl Filter<Extract = impl Reply, Error = Rejection> + Clone {
     path("tree")
         .and(warp::get())
@@ -274,15 +274,15 @@ mod handler {
 
     /// Fetch a [`coco::Blob`].
     pub async fn blob(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
         super::BlobQuery { path, revision }: super::BlobQuery,
     ) -> Result<impl Reply, Rejection> {
         let peer = peer.lock().await;
         let urn = project_urn.parse().map_err(Error::from)?;
-        let project = peer.get_project(&urn)?;
+        let project = coco::get_project(&peer, &urn)?;
         let default_branch = project.default_branch();
-        let blob = peer.with_browser(&urn, |mut browser| {
+        let blob = coco::with_browser(&peer, &urn, |mut browser| {
             coco::blob(&mut browser, default_branch, revision, &path)
         })?;
 
@@ -291,39 +291,41 @@ mod handler {
 
     /// Fetch the list [`coco::Branch`].
     pub async fn branches(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
     ) -> Result<impl Reply, Rejection> {
         let urn = project_urn.parse().map_err(Error::from)?;
         let peer = peer.lock().await;
-        let branches = peer.with_browser(&urn, |browser| coco::branches(browser))?;
+        let branches = coco::with_browser(&peer, &urn, |browser| coco::branches(browser))?;
 
         Ok(reply::json(&branches))
     }
 
     /// Fetch a [`coco::Commit`].
     pub async fn commit(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
         sha1: String,
     ) -> Result<impl Reply, Rejection> {
         let urn = project_urn.parse().map_err(Error::from)?;
         let peer = peer.lock().await;
-        let commit = peer.with_browser(&urn, |mut browser| coco::commit(&mut browser, &sha1))?;
+        let commit =
+            coco::with_browser(&peer, &urn, |mut browser| coco::commit(&mut browser, &sha1))?;
 
         Ok(reply::json(&commit))
     }
 
     /// Fetch the list of [`coco::Commit`] from a branch.
     pub async fn commits(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
         super::CommitsQuery { branch }: super::CommitsQuery,
     ) -> Result<impl Reply, Rejection> {
         let urn = project_urn.parse().map_err(Error::from)?;
         let peer = peer.lock().await;
-        let commits =
-            peer.with_browser(&urn, |mut browser| coco::commits(&mut browser, &branch))?;
+        let commits = coco::with_browser(&peer, &urn, |mut browser| {
+            coco::commits(&mut browser, &branch)
+        })?;
 
         Ok(reply::json(&commits))
     }
@@ -337,12 +339,12 @@ mod handler {
 
     /// Fetch the list [`coco::Branch`] and [`coco::Tag`].
     pub async fn revisions(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
     ) -> Result<impl Reply, Rejection> {
         let urn = project_urn.parse().map_err(Error::from)?;
         let peer = peer.lock().await;
-        let (branches, tags) = peer.with_browser(&urn, |browser| {
+        let (branches, tags) = coco::with_browser(&peer, &urn, |browser| {
             Ok((coco::branches(browser)?, coco::tags(browser)?))
         })?;
 
@@ -376,27 +378,27 @@ mod handler {
 
     /// Fetch the list [`coco::Tag`].
     pub async fn tags(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
     ) -> Result<impl Reply, Rejection> {
         let urn = project_urn.parse().map_err(Error::from)?;
         let peer = peer.lock().await;
-        let tags = peer.with_browser(&urn, |browser| coco::tags(browser))?;
+        let tags = coco::with_browser(&peer, &urn, |browser| coco::tags(browser))?;
 
         Ok(reply::json(&tags))
     }
 
     /// Fetch a [`coco::Tree`].
     pub async fn tree(
-        peer: Arc<Mutex<coco::Peer>>,
+        peer: Arc<Mutex<coco::PeerApi>>,
         project_urn: String,
         super::TreeQuery { prefix, revision }: super::TreeQuery,
     ) -> Result<impl Reply, Rejection> {
         let urn = project_urn.parse().map_err(Error::from)?;
         let peer = peer.lock().await;
-        let project = peer.get_project(&urn)?;
+        let project = coco::get_project(&peer, &urn)?;
         let default_branch = project.default_branch();
-        let tree = peer.with_browser(&urn, |mut browser| {
+        let tree = coco::with_browser(&peer, &urn, |mut browser| {
             coco::tree(&mut browser, default_branch, revision, prefix)
         })?;
 
@@ -773,18 +775,22 @@ mod test {
     async fn blob() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await?;
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = Arc::new(Mutex::new(coco::create_peer_api(config).await?));
+        let owner = coco::fake_owner(peer.lock().await.key().clone()).await;
+        let platinum_project = coco::replicate_platinum(
+            &*peer.lock().await,
+            &owner,
+            "git-platinum",
+            "fixture data",
+            "master",
+        )?;
         let urn = platinum_project.urn();
 
         let revision = "master";
         let default_branch = "master".to_string(); // TODO(finto): need to change this
         let path = "text/arrows.txt";
-        let want = peer.with_browser(&urn, |mut browser| {
+        let want = coco::with_browser(&*peer.lock().await, &urn, |mut browser| {
             coco::blob(
                 &mut browser,
                 &default_branch.clone(),
@@ -793,7 +799,7 @@ mod test {
             )
         })?;
 
-        let api = super::filters(Arc::new(Mutex::new(peer.clone())));
+        let api = super::filters(Arc::clone(&peer));
 
         // Get ASCII blob.
         let res = request()
@@ -857,7 +863,7 @@ mod test {
             .reply(&api)
             .await;
 
-        let want = peer.with_browser(&urn, |browser| {
+        let want = coco::with_browser(&*peer.lock().await, &urn, |browser| {
             coco::blob(browser, &default_branch, Some(revision.to_string()), path)
         })?;
 
@@ -899,15 +905,14 @@ mod test {
     async fn branches() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await?;
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
+        let owner = coco::fake_owner(peer.key().clone()).await;
+        let platinum_project =
+            coco::replicate_platinum(&peer, &owner, "git-platinum", "fixture data", "master")?;
         let urn = platinum_project.urn();
 
-        let want = peer.with_browser(&urn, |browser| coco::branches(browser))?;
+        let want = coco::with_browser(&peer, &urn, |browser| coco::branches(browser))?;
 
         let api = super::filters(Arc::new(Mutex::new(peer)));
         let res = request()
@@ -929,17 +934,17 @@ mod test {
     async fn commit() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await?;
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
+        let owner = coco::fake_owner(peer.key().clone()).await;
+        let platinum_project =
+            coco::replicate_platinum(&peer, &owner, "git-platinum", "fixture data", "master")?;
         let urn = platinum_project.urn();
 
         let sha1 = "3873745c8f6ffb45c990eb23b491d4b4b6182f95";
-        let want =
-            peer.with_browser(&urn, |mut browser| coco::commit_header(&mut browser, sha1))?;
+        let want = coco::with_browser(&peer, &urn, |mut browser| {
+            coco::commit_header(&mut browser, sha1)
+        })?;
 
         let api = super::filters(Arc::new(Mutex::new(peer)));
         let res = request()
@@ -978,19 +983,21 @@ mod test {
     async fn commits() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await?;
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
+        let owner = coco::fake_owner(peer.key().clone()).await;
+        let platinum_project =
+            coco::replicate_platinum(&peer, &owner, "git-platinum", "fixture data", "master")?;
         let urn = platinum_project.urn();
 
         let branch = "master";
         let head = "223aaf87d6ea62eef0014857640fd7c8dd0f80b5";
-        let want = peer.with_browser(&urn, |mut browser| coco::commits(&mut browser, branch))?;
-        let head_commit =
-            peer.with_browser(&urn, |mut browser| coco::commit_header(&mut browser, head))?;
+        let want = coco::with_browser(&peer, &urn, |mut browser| {
+            coco::commits(&mut browser, branch)
+        })?;
+        let head_commit = coco::with_browser(&peer, &urn, |mut browser| {
+            coco::commit_header(&mut browser, head)
+        })?;
 
         let api = super::filters(Arc::new(Mutex::new(peer)));
         let res = request()
@@ -1016,8 +1023,8 @@ mod test {
     async fn local_state() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let peer = coco::Peer::new(config).await?;
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
 
         let path = "../fixtures/git-platinum";
         let api = super::filters(Arc::new(Mutex::new(peer)));
@@ -1050,13 +1057,11 @@ mod test {
     async fn revisions() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await
-            .unwrap();
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
+        let owner = coco::fake_owner(peer.key().clone()).await;
+        let platinum_project =
+            coco::replicate_platinum(&peer, &owner, "git-platinum", "fixture data", "master")?;
         let urn = platinum_project.urn();
 
         // TODO(finto): Get the right URN
@@ -1064,7 +1069,7 @@ mod test {
             "rad:git:hwd1yredksthny1hht3bkhtkxakuzfnjxd8dyk364prfkjxe4xpxsww3try".parse()?;
 
         let want = {
-            let (branches, tags) = peer.with_browser(&urn, |browser| {
+            let (branches, tags) = coco::with_browser(&peer, &urn, |browser| {
                 Ok((coco::branches(browser)?, coco::tags(browser)?))
             })?;
 
@@ -1172,16 +1177,14 @@ mod test {
     async fn tags() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await
-            .unwrap();
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
+        let owner = coco::fake_owner(peer.key().clone()).await;
+        let platinum_project =
+            coco::replicate_platinum(&peer, &owner, "git-platinum", "fixture data", "master")?;
         let urn = platinum_project.urn();
 
-        let want = peer.with_browser(&urn, |browser| coco::tags(browser))?;
+        let want = coco::with_browser(&peer, &urn, |browser| coco::tags(browser))?;
 
         let api = super::filters(Arc::new(Mutex::new(peer)));
         let res = request()
@@ -1205,19 +1208,18 @@ mod test {
     async fn tree() -> Result<(), error::Error> {
         let tmp_dir = tempfile::tempdir()?;
         let key = SecretKey::new();
-        let config = coco::default_config(key, tmp_dir)?;
-        let mut peer = coco::Peer::new(config).await?;
-        let owner = coco::fake_owner(&peer).await;
-        let platinum_project = peer
-            .replicate_platinum(&owner, "git-platinum", "fixture data", "master")
-            .await?;
+        let config = coco::config::default(key, tmp_dir)?;
+        let peer = coco::create_peer_api(config).await?;
+        let owner = coco::fake_owner(peer.key().clone()).await;
+        let platinum_project =
+            coco::replicate_platinum(&peer, &owner, "git-platinum", "fixture data", "master")?;
         let urn = platinum_project.urn();
 
         let revision = "master";
         let prefix = "src";
 
         let default_branch = "master".to_string(); // TODO(finto): need to change this
-        let want = peer.with_browser(&urn, |mut browser| {
+        let want = coco::with_browser(&peer, &urn, |mut browser| {
             coco::tree(
                 &mut browser,
                 &default_branch,
